counter

+++ classwork/class10/library/models.py
@@ -3,4 +3,4 @@ from django.db import models
 class Book(models.Model):
     title = models.CharField(max_length=500)
     author = models.TextField()
-    year = models.DecimalField(max_digits=10, decimal_places=2)
+    year = models.IntegerField()
